+++ src/main/java/com/javPOL/magazineJava/dao/ProductDAO/ProductDao.java
@@ -3,5 +3,5 @@ package com.javPOL.magazineJava.dao.ProductDAO;
 import com.javPOL.magazineJava.dao.Dao;
 import com.javPOL.magazineJava.model.Product;
 
-public interface ProductDao extends Dao<Product, Integer> {
+public interface ProductDao extends Dao<Product, Long> {
 }
