+++ src/main/java/com/javPOL/magazineJava/dao/WarehouseDAO/WarehouseDao.java
@@ -3,5 +3,5 @@ package com.javPOL.magazineJava.dao.WarehouseDAO;
 import com.javPOL.magazineJava.dao.Dao;
 import com.javPOL.magazineJava.model.Warehouse;
 
-public interface WarehouseDao extends Dao<Warehouse, Integer> {
+public interface WarehouseDao extends Dao<Warehouse, Long> {
 }
